ft_printf.h: (W) changed the input type for putxnbr and putxxnbr back to unsigned int. It cannot take any other type of input because this is the inly format thats expected from a %x specifier

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dimitris <dimitris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: dchrysov <dchrysov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 09:57:23 by dchrysov          #+#    #+#             */
-/*   Updated: 2024/10/21 22:45:34 by dimitris         ###   ########.fr       */
+/*   Updated: 2024/10/22 13:25:48 by dchrysov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,11 +16,12 @@
 # include <unistd.h>
 # include <stdarg.h>
 # include <limits.h>
+# include <stdlib.h>
 
 int		ft_printf(const char *str, ...);
 int		ft_putunbr_fd(int n, int fd);
 int		ft_putxxnbr_fd(unsigned int n, int fd);
-int		ft_putxnbr_fd(unsigned long n, int fd);
+int		ft_putxnbr_fd(unsigned int n, int fd);
 int		ft_putstr_fd(char *s, int fd);
 int		ft_putnbr_fd(int n, int fd);
 int		ft_putchar_fd(char c, int fd);
